requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
python-dotenv
python-dateutil
humanize
more-itertools
rapidfuzz
colour
Pillow
typing
TagScriptEngine
aiohttp
regex
pyparsing